Removing instance API as it was un-used and simplifying the nature of an instance.
