+++ packages/server/src/api/routes/index.js
@@ -1,7 +1,6 @@
 const authRoutes = require("./auth")
 const pageRoutes = require("./pages")
 const userRoutes = require("./user")
-const instanceRoutes = require("./instance")
 const applicationRoutes = require("./application")
 const tableRoutes = require("./table")
 const rowRoutes = require("./row")
@@ -21,7 +20,6 @@ module.exports = {
   authRoutes,
   pageRoutes,
   userRoutes,
-  instanceRoutes,
   applicationRoutes,
   rowRoutes,
   tableRoutes,
