--- packages/server/src/api/routes/instance.js
@@ -1,12 +1,0 @@
-const Router = require("@koa/router")
-const controller = require("../controllers/instance")
-const authorized = require("../../middleware/authorized")
-const { BUILDER } = require("../../utilities/accessLevels")
-
-const router = Router()
-
-router
-  .post("/api/instances", authorized(BUILDER), controller.create)
-  .delete("/api/instances/:instanceId", authorized(BUILDER), controller.destroy)
-
-module.exports = router
