+++ packages/server/src/api/routes/tests/automation.spec.js
@@ -49,7 +49,7 @@ describe("/automations", () => {
 
   beforeEach(async () => {
     app = await createApplication(request)
-    instanceId = app.instances[0]._id
+    instanceId = app.instance._id
     if (automation) await destroyDocument(automation.id)
   })
