+++ packages/server/src/api/routes/tests/user.spec.js
@@ -23,7 +23,7 @@ describe("/users", () => {
 
   beforeEach(async () => {
     app = await createApplication(request)
-    instanceId = app.instances[0]._id
+    instanceId = app.instance._id
   });
 
   afterAll(() => {
