+++ packages/server/src/api/routes/tests/view.spec.js
@@ -39,7 +39,7 @@ describe("/views", () => {
 
   beforeEach(async () => {
     app = await createApplication(request)
-    instanceId = app.instances[0]._id
+    instanceId = app.instance._id
   })
 
   afterAll(() => {
